+++ tests/model_serving/model_server/storage/pvc/test_kserve_pvc_rwx.py
@@ -1,6 +1,6 @@
 import shlex
 from typing import List
-from tests.constants import KServeDeploymentType
+from utilities.constants import KServeDeploymentType
 
 import pytest
